rename main game executive file and update imports
